refactor(network): reorganize upstream dialing and proxy support
- Split upstream construction, config parsing, resolver logic, dialing, and protocol connections into focused modules.
- Move SOCKS5 option parsing into shared network proxy support for HTTP downloads, requests, upgrades, and upstreams.
- Keep connection pooling separate from protocol connection implementations, with DoH helpers scoped under connection code.
- Validate the refactor with cargo check, upstream/forward tests, feature gating, and the full cargo test suite.
