+++ src/network/upstream/mod.rs
@@ -6,2098 +6,22 @@
 //! This module builds outbound resolvers used by forwarding-style executors.
 //! It turns upstream configuration into protocol-specific clients with shared
 //! pooling, bootstrap resolution, timeout handling, and fallback behavior.
-//!
-//! # Supported Protocols
-//! - **UDP**: Standard DNS over UDP (port 53)
-//! - **TCP**: DNS over TCP (port 53) with pipelining support
-//! - **DoT**: DNS over TLS (port 853)
-//! - **DoQ**: DNS over QUIC (port 853)
-//! - **DoH**: DNS over HTTPS via HTTP/2 or HTTP/3 (port 443)
-//!
-//! # Connection Management
-//! - **Pipeline Pool**: Multiple concurrent requests per connection
-//! - **Reuse Pool**: Connection recycling with idle timeout
-//! - **Bootstrap**: Efficient hostname resolution for upstream servers
-//! - **Fallback**: UDP → TCP fallback for truncated responses
-//!
-//! # Performance Features
-//! - Lock-free connection pooling
-//! - Automatic connection scaling
-//! - Request pipelining for TCP/TLS
-//! - Connection reuse with idle management
-//! - Zero-copy DNS message handling where possible
-//!
-//! This code sits on the outbound edge of the request pipeline. It should keep
-//! network concerns isolated from policy composition while remaining explicit
-//! about connection reuse, timeout boundaries, and protocol semantics.
-
-use std::fmt::Debug;
-use std::net::{IpAddr, SocketAddr};
-use std::str::FromStr;
-use std::sync::Arc;
-use std::time::Duration;
-
-use arc_swap::ArcSwap;
-use async_trait::async_trait;
-use serde::Deserialize;
-#[cfg(feature = "_http-client")]
-use tokio::net::TcpStream;
-use tokio::sync::Mutex;
-use tracing::{debug, info, warn};
-use url::Url;
-
-use crate::core::error::{DnsError, Result};
-use crate::core::system_utils::deserialize_duration_option;
-use crate::network::upstream::bootstrap::Bootstrap;
-#[cfg(feature = "upstream-doh")]
-use crate::network::upstream::pool::conn_h2::{H2Connection, H2ConnectionBuilder};
-#[cfg(feature = "upstream-doh3")]
-use crate::network::upstream::pool::conn_h3::{H3Connection, H3ConnectionBuilder};
-#[cfg(feature = "upstream-doq")]
-use crate::network::upstream::pool::conn_quic::{QuicConnection, QuicConnectionBuilder};
-use crate::network::upstream::pool::conn_tcp::{TcpConnection, TcpConnectionBuilder};
-use crate::network::upstream::pool::conn_udp::{UdpConnection, UdpConnectionBuilder};
-use crate::network::upstream::pool::pool_pipeline::PipelinePool;
-use crate::network::upstream::pool::pool_reuse::ReusePool;
-use crate::network::upstream::pool::{
-    Connection, ConnectionBuilder, ConnectionPool, DeadlineOutcome, QueryTimeoutPolicy,
-};
-use crate::network::upstream::utils::try_lookup_server_name;
-use crate::proto::Message;
 
 mod bootstrap;
+mod builder;
+mod config;
+mod conn;
+mod dial;
 mod pool;
-mod utils;
+mod resolver;
 
-pub use pool::QueryDeadline;
-
-/// Supported upstream connection types
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub enum ConnectionType {
-    UDP,
-    TCP,
-    DoT,
-    DoQ,
-    DoH,
-}
-
-#[allow(unused)]
-impl ConnectionType {
-    /// Returns the default port for each connection type
-    pub fn default_port(&self) -> u16 {
-        match self {
-            ConnectionType::UDP => 53,
-            ConnectionType::TCP => 53,
-            ConnectionType::DoT => 853,
-            ConnectionType::DoQ => 853,
-            ConnectionType::DoH => 443,
-        }
-    }
-
-    /// Returns all supported URL schemes for this connection type
-    pub fn schemes(&self) -> Vec<&str> {
-        match self {
-            ConnectionType::UDP => vec!["udp", ""],
-            ConnectionType::TCP => vec!["tcp", "tcp+pipeline"],
-            ConnectionType::DoT => vec!["tls", "tls+pipeline"],
-            ConnectionType::DoQ => vec!["doq", "quic"],
-            ConnectionType::DoH => vec!["doh", "https", "h3"],
-        }
-    }
-}
-
-/// Configuration for building an upstream DNS server connection
-///
-/// This structure is typically deserialized from YAML/JSON configuration files
-/// and contains all parameters needed to establish a connection to an upstream
-/// DNS server.
-///
-/// # Examples
-///
-/// Basic UDP configuration:
-/// ```yaml
-/// addr: "8.8.8.8:53"
-/// ```
-///
-/// DoH with bootstrap:
-/// ```yaml
-/// addr: "https://dns.google.com/dns-query"
-/// bootstrap: "8.8.8.8:53"
-/// timeout: 5s
-/// ```
-#[derive(Deserialize, Debug, Clone)]
-pub struct UpstreamConfig {
-    /// Optional tag for identifying this upstream in logs
-    pub tag: Option<String>,
-
-    /// DNS server address in URL format
-    ///
-    /// Supported formats:
-    /// - `udp://8.8.8.8:53` or `8.8.8.8` - DNS over UDP
-    /// - `tcp://8.8.8.8:53` - DNS over TCP
-    /// - `tls://dns.google.com:853` - DNS over TLS (DoT)
-    /// - `quic://dns.adguard.com:853` - DNS over QUIC (DoQ)
-    /// - `https://dns.google.com/dns-query` - DNS over HTTPS (DoH)
-    pub addr: String,
-
-    /// Direct IP address to use for connection (bypasses DNS resolution)
-    ///
-    /// Useful when you want to connect to a specific IP but use SNI for TLS.
-    /// If provided, this IP is used instead of resolving the hostname from
-    /// `addr`. Mutually exclusive with `bootstrap` at runtime: when both are
-    /// configured, `dial_addr` takes precedence and `bootstrap` is ignored.
-    pub dial_addr: Option<IpAddr>,
-
-    /// Override the server port (if not specified in `addr`)
-    ///
-    /// Defaults to protocol-specific standard ports if not provided:
-    /// - UDP/TCP: 53
-    /// - DoT/DoQ: 853
-    /// - DoH: 443
-    pub port: Option<u16>,
-
-    /// Bootstrap DNS server for resolving the upstream hostname
-    ///
-    /// Recommended when `addr` contains a hostname instead of an IP address.
-    /// Without bootstrap, hostname resolution is deferred to connection time
-    /// and uses the operating system resolver. The bootstrap server must be
-    /// specified as IP:port (e.g., "8.8.8.8:53") to avoid circular
-    /// dependencies in DNS resolution; hostnames are rejected. Mutually
-    /// exclusive with `dial_addr` at runtime: when both are configured,
-    /// `dial_addr` takes precedence and bootstrap resolution is skipped.
-    ///
-    /// # Example
-    /// ```yaml
-    /// addr: "https://dns.google.com/dns-query"
-    /// bootstrap: "8.8.8.8:53"  # Use Google's IP to resolve dns.google.com
-    /// ```
-    pub bootstrap: Option<String>,
-
-    /// IP version preference for bootstrap DNS resolution
-    ///
-    /// - `Some(4)` or `None`: Resolve to IPv4 (A records)
-    /// - `Some(6)`: Resolve to IPv6 (AAAA records)
-    pub bootstrap_version: Option<u8>,
-
-    /// SOCKS5 proxy server for upstream connections
-    ///
-    /// When specified, all DNS connections to the upstream server will be
-    /// routed through this SOCKS5 proxy. The proxy address can be either an
-    /// IP address or a hostname (which will be resolved using system DNS).
-    ///
-    /// Supports two formats:
-    /// - **Without authentication**: `"host:port"`
-    ///   - Example: `"127.0.0.1:1080"`
-    ///   - Example: `"proxy.example.com:1080"`
-    ///
-    /// - **With authentication**: `"username:password@host:port"`
-    ///   - Example: `"user:pass@127.0.0.1:1080"`
-    ///   - Example: `"myuser:mypass@proxy.example.com:1080"`
-    ///
-    /// **Note**: If the proxy hostname fails to resolve, the upstream will
-    /// not be created and an error will be logged during initialization.
-    ///
-    /// # IPv6 Support
-    /// IPv6 addresses must be enclosed in brackets:
-    /// - `"[::1]:1080"` - IPv6 without auth
-    /// - `"user:pass@[2001:db8::1]:1080"` - IPv6 with auth
-    pub socks5: Option<String>,
-
-    /// Connection idle timeout in seconds
-    ///
-    /// Used by connection pools to recycle idle connections and bound
-    /// long-lived unused sockets.
-    ///
-    /// The value accepts a duration string or a number. When a bare number is
-    /// provided, it is interpreted as seconds.
-    ///
-    /// Examples:
-    /// - `"5s"`
-    /// - `"5"` // equivalent to `"5s"`
-    #[serde(default, deserialize_with = "deserialize_duration_option")]
-    pub idle_timeout: Option<Duration>,
-
-    /// Maximum number of connections in the pool
-    ///
-    /// Used as the pool size upper bound to limit per-upstream resource usage.
-    pub max_conns: Option<usize>,
-
-    /// Minimum number of connections to keep warm in the pool
-    ///
-    /// Defaults to 0, which preserves lazy connection creation.
-    pub min_conns: Option<usize>,
-
-    /// Skip TLS certificate verification (**INSECURE**, testing only!)
-    ///
-    /// When `true`, disables certificate validation for TLS/QUIC/DoH
-    /// connections. **Security Warning**: This makes connections vulnerable
-    /// to MITM attacks. Only use for testing or with self-signed
-    /// certificates you trust.
-    pub insecure_skip_verify: Option<bool>,
-
-    /// DNS query timeout duration
-    ///
-    /// Maximum time to wait for a DNS response before considering the query
-    /// failed.
-    ///
-    /// The value accepts a duration string or a number. When a bare number is
-    /// provided, it is interpreted as seconds.
-    ///
-    /// Defaults to 5 seconds if not specified.
-    ///
-    /// Examples:
-    /// - `"5s"`
-    /// - `"5"` // equivalent to `"5s"`
-    #[serde(default, deserialize_with = "deserialize_duration_option")]
-    pub timeout: Option<Duration>,
-
-    /// Enable request pipelining for TCP/DoT connections
-    ///
-    /// When `true`, allows multiple concurrent queries over a single TCP
-    /// connection. When `false`, uses connection pooling with one query per
-    /// connection. Only applicable to TCP and DoT protocols.
-    pub enable_pipeline: Option<bool>,
-
-    /// Enable HTTP/3 for DoH connections
-    ///
-    /// When `true`, uses HTTP/3 (QUIC) instead of HTTP/2 for DoH.
-    /// Requires the upstream server to support HTTP/3.
-    pub enable_http3: Option<bool>,
-
-    /// Linux SO_MARK socket option for policy routing
-    ///
-    /// Sets the mark on outgoing packets, which can be used with
-    /// iptables/nftables for advanced routing policies.
-    /// **Linux only** - ignored on other platforms.
-    pub so_mark: Option<u32>,
-
-    /// Linux SO_BINDTODEVICE - bind socket to specific network interface
-    ///
-    /// Forces the socket to use a specific network interface (e.g., "eth0",
-    /// "wlan0"). Useful for multi-homed systems or VPN scenarios.
-    /// **Linux only** - ignored on other platforms.
-    pub bind_to_device: Option<String>,
-}
-
-#[async_trait]
-#[allow(unused)]
-pub trait Upstream: Send + Sync + Debug {
-    /// **Internal API - Do not call directly!**
-    ///
-    /// Send a DNS query using the provided end-to-end query deadline.
-    ///
-    /// # For Implementors
-    /// Implement this method to provide the actual DNS query logic.
-    ///
-    /// # For Callers
-    /// **Always use `query()` or `query_with_deadline()` instead!**
-    #[doc(hidden)]
-    async fn inner_query(&self, request: Message, deadline: QueryDeadline) -> Result<Message>;
-
-    /// Return the connection configuration information
-    ///
-    /// Provides access to all upstream connection parameters including
-    /// connection type, timeout, addresses, and protocol-specific settings.
-    fn connection_info(&self) -> &ConnectionInfo;
-
-    /// Return the timeout duration for this upstream
-    ///
-    /// Default implementation reads from connection_info.
-    /// Can be overridden if custom timeout logic is needed.
-    #[inline]
-    fn timeout(&self) -> Duration {
-        self.connection_info().timeout
-    }
-
-    /// Return the connection type of this upstream
-    ///
-    /// Convenience method for accessing connection_info.connection_type.
-    #[inline]
-    fn connection_type(&self) -> ConnectionType {
-        self.connection_info().connection_type
-    }
-
-    /// Whether `inner_query` owns deadline enforcement and timeout cleanup.
-    ///
-    /// Pool-backed implementations must return `true` so the pool can observe
-    /// deadline expiry and apply its connection retirement/close policy.
-    #[inline]
-    fn handles_query_deadline(&self) -> bool {
-        false
-    }
-
-    /// Send a DNS query with an existing upstream deadline.
-    async fn query_with_deadline(
-        &self,
-        message: Message,
-        deadline: QueryDeadline,
-    ) -> Result<Message> {
-        if deadline.remaining().is_none() {
-            warn!(
-                timeout_secs = self.timeout().as_secs_f64(),
-                "Upstream DNS query timeout"
-            );
-            return Err(deadline.timeout_error());
-        }
-        if self.handles_query_deadline() {
-            return self.inner_query(message, deadline).await;
-        }
-        match deadline.run(self.inner_query(message, deadline)).await {
-            DeadlineOutcome::Completed(result) => result,
-            DeadlineOutcome::Expired => {
-                warn!(
-                    timeout_secs = self.timeout().as_secs_f64(),
-                    "Upstream DNS query timeout"
-                );
-                Err(deadline.timeout_error())
-            }
-        }
-    }
-
-    /// Send a DNS query with unified deadline handling
-    ///
-    /// This is the **recommended API** for all DNS queries.
-    /// Automatically applies timeout based on `timeout()` configuration.
-    ///
-    /// # Performance Notes
-    /// - Message is moved (not cloned) to avoid allocation overhead
-    /// - Timeout error logging uses structured fields for zero-copy
-    /// - Only logs on timeout, not on successful queries (hot path
-    ///   optimization)
-    ///
-    /// # Errors
-    /// - Returns `DnsError::plugin` on timeout
-    /// - Returns upstream-specific errors on query failures
-    async fn query(&self, message: Message) -> Result<Message> {
-        let deadline = QueryDeadline::new(self.timeout());
-        self.query_with_deadline(message, deadline).await
-    }
-}
-
-/// SOCKS5 proxy configuration with resolved socket address
-///
-/// This struct contains the parsed and resolved SOCKS5 proxy information,
-/// ready to be used for establishing proxy connections.
-///
-/// # Fields
-/// - `username`: Optional SOCKS5 authentication username
-/// - `password`: Optional SOCKS5 authentication password
-/// - `socket_addr`: Resolved proxy server socket address (IP + port)
-///
-/// # Note
-/// The hostname in the original configuration (if any) has already been
-/// resolved to an IP address when this struct is created.
-#[derive(Debug, Clone)]
-#[allow(dead_code)]
-pub struct Socks5Opt {
-    username: Option<String>,
-    password: Option<String>,
-    socket_addr: SocketAddr,
-}
-
-/// Runtime connection information for upstream DNS servers
-///
-/// Parsed and processed configuration ready for connection establishment.
-/// Created from `UpstreamConfig` via `From` trait, passed to connection
-/// builders.
-///
-/// Thread-safe (`Clone`) for sharing across multiple connection instances.
-#[derive(Debug, Clone)]
-#[allow(unused)]
-pub struct ConnectionInfo {
-    /// Optional tag for identifying this upstream in logs
-    pub tag: Option<String>,
-
-    /// Protocol type (auto-detected from URL scheme: udp://, tcp://, tls://, quic://, https://)
-    pub connection_type: ConnectionType,
-
-    /// Original address string from configuration (for logging)
-    pub raw_addr: String,
-
-    /// Literal or explicitly configured IP address (`None` if hostname
-    /// resolution is deferred to bootstrap or connection time)
-    pub remote_ip: Option<IpAddr>,
-
-    /// Server port (protocol default or explicitly configured)
-    pub port: u16,
-
-    /// SOCKS5 proxy configuration
-    pub socks5: Option<Socks5Opt>,
-
-    /// Bootstrap resolver for dynamic hostname resolution with TTL caching
-    bootstrap: Option<Arc<Bootstrap>>,
-
-    /// DoH request path (e.g., `/dns-query`), empty for non-HTTP protocols
-    pub path: String,
-
-    /// Server hostname for TLS SNI and certificate validation
-    pub server_name: String,
-
-    /// Skip TLS certificate verification (**INSECURE** - testing only)
-    pub insecure_skip_verify: bool,
-
-    /// Connection idle timeout in seconds
-    pub idle_timeout: Duration,
-
-    /// Maximum number of connections in the pool
-    pub max_conns: Option<usize>,
-
-    /// Minimum number of connections to keep warm in the pool
-    pub min_conns: Option<usize>,
-
-    /// DNS query timeout (includes I/O, handshakes, and round-trip time)
-    pub timeout: Duration,
-
-    /// Request pipelining for TCP/DoT (`None` = protocol default)
-    pub enable_pipeline: Option<bool>,
-
-    /// Use HTTP/3 (true) instead of HTTP/2 (false) for DoH
-    pub enable_http3: bool,
-
-    /// Linux SO_MARK for packet marking (policy routing)
-    pub so_mark: Option<u32>,
-
-    /// Linux SO_BINDTODEVICE - bind to specific network interface
-    pub bind_to_device: Option<String>,
-}
-
-impl ConnectionInfo {
-    const DEFAULT_CONN_IDLE_TIME: Duration = Duration::from_secs(10);
-    const DEFAULT_MAX_CONNS_LOAD: u16 = 64;
-    const DEFAULT_MAX_CONNS_SIZE: usize = 64;
-    const DEFAULT_MIN_CONNS_SIZE: usize = 0;
-    const DEFAULT_QUERY_TIMEOUT: Duration = Duration::from_secs(5);
-    const MAX_CONFIGURED_CONNS_SIZE: usize = 4096;
-
-    pub fn with_addr(addr: &str) -> Result<Self> {
-        let (connection_type, host, port, path, _) = detect_connection_type(addr)?;
-        let port = port.unwrap_or(connection_type.default_port());
-
-        debug!(
-            "Building ConnectionInfo: type={:?}, host={}, port={}, path={}",
-            connection_type, host, port, path
-        );
-
-        let remote_ip = static_remote_ip_from_host(&host, None);
-
-        Ok(ConnectionInfo {
-            tag: None,
-            remote_ip,
-            port,
-            socks5: None,
-            connection_type,
-            bootstrap: None,
-            path,
-            timeout: Self::DEFAULT_QUERY_TIMEOUT,
-            server_name: host,
-            insecure_skip_verify: false,
-            idle_timeout: Self::DEFAULT_CONN_IDLE_TIME,
-            raw_addr: addr.to_string(),
-            enable_pipeline: None,
-            enable_http3: false,
-            so_mark: None,
-            bind_to_device: None,
-            max_conns: None,
-            min_conns: None,
-        })
-    }
-
-    pub fn validate_addr(addr: &str) -> Result<()> {
-        detect_connection_type(addr).map(|_| ())
-    }
-
-    fn max_conns_or_default(&self) -> usize {
-        self.max_conns.unwrap_or(Self::DEFAULT_MAX_CONNS_SIZE)
-    }
-
-    fn min_conns_or_default(&self) -> usize {
-        self.min_conns.unwrap_or(Self::DEFAULT_MIN_CONNS_SIZE)
-    }
-}
-
-impl TryFrom<UpstreamConfig> for ConnectionInfo {
-    type Error = DnsError;
-
-    fn try_from(upstream_config: UpstreamConfig) -> Result<Self> {
-        let UpstreamConfig {
-            tag,
-            addr,
-            dial_addr,
-            port: config_port,
-            bootstrap,
-            bootstrap_version,
-            socks5,
-            idle_timeout,
-            max_conns,
-            min_conns,
-            insecure_skip_verify,
-            timeout,
-            enable_pipeline,
-            enable_http3,
-            so_mark,
-            bind_to_device,
-        } = upstream_config;
-        let (connection_type, host, port, path, helper_flags) = detect_connection_type(&addr)?;
-        let enable_pipeline = if helper_flags.force_pipeline {
-            Some(true)
-        } else {
-            enable_pipeline
-        };
-        let enable_http3 = if helper_flags.force_http3 {
-            true
-        } else {
-            enable_http3.unwrap_or(false)
-        };
-        let port = config_port
-            .or(port)
-            .unwrap_or(connection_type.default_port());
-
-        if let Some(max_conns) = max_conns {
-            if max_conns == 0 {
-                return Err(DnsError::plugin(
-                    "upstream max_conns must be greater than 0",
-                ));
-            }
-            if max_conns > ConnectionInfo::MAX_CONFIGURED_CONNS_SIZE {
-                return Err(DnsError::plugin(format!(
-                    "upstream max_conns must be <= {}",
-                    ConnectionInfo::MAX_CONFIGURED_CONNS_SIZE
-                )));
-            }
-        }
-        if let Some(min_conns) = min_conns {
-            if min_conns > ConnectionInfo::MAX_CONFIGURED_CONNS_SIZE {
-                return Err(DnsError::plugin(format!(
-                    "upstream min_conns must be <= {}",
-                    ConnectionInfo::MAX_CONFIGURED_CONNS_SIZE
-                )));
-            }
-
-            let effective_max_conns = max_conns.unwrap_or(ConnectionInfo::DEFAULT_MAX_CONNS_SIZE);
-            if min_conns > effective_max_conns {
-                return Err(DnsError::plugin(format!(
-                    "upstream min_conns must be <= max_conns (effective max_conns: {})",
-                    effective_max_conns
-                )));
-            }
-        }
-        if !matches!(bootstrap_version, None | Some(4) | Some(6)) {
-            return Err(DnsError::plugin(
-                "upstream bootstrap_version must be 4 or 6",
-            ));
-        }
-
-        debug!(
-            "Building ConnectionInfo: type={:?}, host={}, port={}, path={}",
-            connection_type, &host, port, path
-        );
-
-        let dial_addr_configured = dial_addr.is_some();
-        let remote_ip = static_remote_ip_from_host(&host, dial_addr);
-
-        if dial_addr_configured && bootstrap.is_some() {
-            warn!(
-                upstream = %addr,
-                "Both dial_addr and bootstrap are configured; dial_addr takes precedence and bootstrap will be ignored"
-            );
-        }
-
-        let bootstrap = if let Some(bootstrap_server) = bootstrap
-            && remote_ip.is_none()
-        {
-            Some(Arc::new(Bootstrap::new(
-                &bootstrap_server,
-                &host,
-                bootstrap_version,
-            )?))
-        } else {
-            None
-        };
-
-        let socks5 = if let Some(socks5_str) = socks5 {
-            match connection_type {
-                ConnectionType::TCP | ConnectionType::DoT => parse_socks5_opt(&socks5_str),
-                ConnectionType::DoH => {
-                    if enable_http3 {
-                        warn!("Sock5 proxy only support tcp portal");
-                        None
-                    } else {
-                        parse_socks5_opt(&socks5_str)
-                    }
-                }
-                _ => {
-                    warn!("Sock5 proxy only support tcp portal");
-                    None
-                }
-            }
-        } else {
-            None
-        };
-
-        Ok(ConnectionInfo {
-            tag,
-            remote_ip,
-            port,
-            socks5,
-            connection_type,
-            bootstrap,
-            path,
-            timeout: timeout.unwrap_or(Self::DEFAULT_QUERY_TIMEOUT),
-            server_name: host,
-            insecure_skip_verify: insecure_skip_verify.unwrap_or(false),
-            idle_timeout: idle_timeout.unwrap_or(Self::DEFAULT_CONN_IDLE_TIME),
-            raw_addr: addr,
-            enable_pipeline,
-            enable_http3,
-            so_mark,
-            bind_to_device,
-            max_conns,
-            min_conns,
-        })
-    }
-}
-
-/// Determine the startup-known remote IP address.
-///
-/// # Arguments
-/// - `host`: The hostname or IP address string
-/// - `dial_addr`: Optional pre-configured IP address to use directly
-///
-/// # Returns
-/// `Some(IpAddr)` if an IP address is explicitly configured or present
-/// literally in `host`; `None` for hostnames. Hostname resolution is deferred
-/// to bootstrap or connection creation so startup and config validation do not
-/// depend on the local system resolver.
-fn static_remote_ip_from_host(host: &str, dial_addr: Option<IpAddr>) -> Option<IpAddr> {
-    // 1. Use dial_addr if provided
-    if let Some(ip) = dial_addr {
-        return Some(ip);
-    }
-
-    // 2. Try parsing as IP address
-    if let Ok(ip) = IpAddr::from_str(host) {
-        return Some(ip);
-    }
-
-    None
-}
-
-/// Detect the connection type from the config address
-#[derive(Clone, Copy, Debug, Default)]
-struct HelperFlags {
-    force_pipeline: bool,
-    force_http3: bool,
-}
-
-fn detect_connection_type(
-    addr: &str,
-) -> Result<(ConnectionType, String, Option<u16>, String, HelperFlags)> {
-    if !addr.contains("//") {
-        return detect_connection_type(&("udp://".to_owned() + addr));
-    }
-
-    let url =
-        Url::parse(addr).map_err(|e| DnsError::plugin(format!("invalid upstream URL: {}", e)))?;
-    let mut helper_flags = HelperFlags::default();
-    let connection_type;
-
-    let host = url
-        .host_str()
-        .map(|host| host.to_owned())
-        .ok_or_else(|| DnsError::plugin("invalid upstream URL: no host specified"))?;
-
-    match url.scheme() {
-        "udp" => {
-            connection_type = ConnectionType::UDP;
-        }
-        "tcp" => {
-            connection_type = ConnectionType::TCP;
-        }
-        "tcp+pipeline" => {
-            helper_flags.force_pipeline = true;
-            connection_type = ConnectionType::TCP;
-        }
-        "tls" => {
-            connection_type = ConnectionType::DoT;
-        }
-        "tls+pipeline" => {
-            helper_flags.force_pipeline = true;
-            connection_type = ConnectionType::DoT;
-        }
-        "quic" | "doq" => {
-            connection_type = ConnectionType::DoQ;
-        }
-        "https" | "doh" => {
-            connection_type = ConnectionType::DoH;
-        }
-        "h3" => {
-            helper_flags.force_http3 = true;
-            connection_type = ConnectionType::DoH;
-        }
-        other => {
-            return Err(DnsError::plugin(format!(
-                "invalid upstream URL scheme: {}",
-                other
-            )));
-        }
-    };
-
-    debug!(
-        "Detected upstream: scheme={}, type={:?}, host={}, port={:?}, path={}",
-        url.scheme(),
-        connection_type,
-        host,
-        url.port(),
-        url.path()
-    );
-
-    Ok((
-        connection_type,
-        host,
-        url.port(),
-        url.path().to_string(),
-        helper_flags,
-    ))
-}
-
-/// Builder for creating upstream instances
-pub struct UpstreamBuilder;
-
-impl UpstreamBuilder {
-    pub fn with_connection_info(connection_info: ConnectionInfo) -> Result<Box<dyn Upstream>> {
-        debug!(
-            "Creating upstream: type={:?}, remote={:?}, port={}",
-            connection_info.connection_type, connection_info.remote_ip, connection_info.port
-        );
-
-        if connection_info.bootstrap.is_none() {
-            let upstream: Box<dyn Upstream> = match connection_info.connection_type {
-                ConnectionType::UDP => {
-                    debug!("Creating UDP upstream for {}", connection_info.raw_addr);
-                    let builder = UdpConnectionBuilder::new(
-                        &connection_info,
-                        pipeline_request_map_capacity(),
-                    );
-                    let main_pool = PipelinePool::new(
-                        main_pool_min_conns(&connection_info),
-                        connection_info.max_conns_or_default(),
-                        ConnectionInfo::DEFAULT_MAX_CONNS_LOAD,
-                        connection_info.idle_timeout,
-                        Box::new(builder),
-                        QueryTimeoutPolicy::Reuse,
-                        connection_info.timeout,
-                    );
-
-                    let tcp_builder =
-                        TcpConnectionBuilder::new(&connection_info, reuse_request_map_capacity());
-                    let fallback_pool = ReusePool::new(
-                        udp_truncated_fallback_min_conns(),
-                        connection_info.max_conns_or_default(),
-                        connection_info.idle_timeout,
-                        Box::new(tcp_builder),
-                        QueryTimeoutPolicy::Close,
-                        connection_info.timeout,
-                    );
-
-                    Box::new(UdpTruncatedUpstream {
-                        connection_info,
-                        main_pool,
-                        fallback_pool,
-                    })
-                }
-                ConnectionType::TCP => {
-                    debug!("Creating TCP upstream for {}", connection_info.raw_addr);
-                    if connection_info.enable_pipeline.unwrap_or(false) {
-                        let builder = TcpConnectionBuilder::new(
-                            &connection_info,
-                            pipeline_request_map_capacity(),
-                        );
-                        Box::new(create_pipeline_pool(connection_info, Box::new(builder)))
-                    } else {
-                        let builder = TcpConnectionBuilder::new(
-                            &connection_info,
-                            reuse_request_map_capacity(),
-                        );
-                        Box::new(create_reuse_pool(connection_info, Box::new(builder)))
-                    }
-                }
-                #[cfg(feature = "upstream-dot")]
-                ConnectionType::DoT => {
-                    debug!("Creating DoT upstream for {}", connection_info.raw_addr);
-                    if connection_info.enable_pipeline.unwrap_or(false) {
-                        let builder = TcpConnectionBuilder::new(
-                            &connection_info,
-                            pipeline_request_map_capacity(),
-                        );
-                        Box::new(create_pipeline_pool(connection_info, Box::new(builder)))
-                    } else {
-                        let builder = TcpConnectionBuilder::new(
-                            &connection_info,
-                            reuse_request_map_capacity(),
-                        );
-                        Box::new(create_reuse_pool(connection_info, Box::new(builder)))
-                    }
-                }
-                #[cfg(not(feature = "upstream-dot"))]
-                ConnectionType::DoT => {
-                    return Err(DnsError::plugin(
-                        "upstream DoT is not compiled into this build; \
-                         rebuild with --features upstream-dot",
-                    ));
-                }
-                #[cfg(feature = "upstream-doq")]
-                ConnectionType::DoQ => {
-                    debug!("Creating QUIC upstream for {}", connection_info.raw_addr);
-                    let builder = QuicConnectionBuilder::new(&connection_info);
-                    Box::new(create_pipeline_pool(connection_info, Box::new(builder)))
-                }
-                #[cfg(not(feature = "upstream-doq"))]
-                ConnectionType::DoQ => {
-                    return Err(DnsError::plugin(
-                        "upstream DoQ is not compiled into this build; \
-                         rebuild with --features upstream-doq",
-                    ));
-                }
-                #[cfg(feature = "upstream-doh")]
-                ConnectionType::DoH => {
-                    debug!(
-                        "Creating DoH upstream for {} (HTTP/{})",
-                        connection_info.raw_addr,
-                        if connection_info.enable_http3 {
-                            "3"
-                        } else {
-                            "2"
-                        }
-                    );
-                    if connection_info.enable_http3 {
-                        #[cfg(feature = "upstream-doh3")]
-                        {
-                            let builder = H3ConnectionBuilder::new(&connection_info);
-                            Box::new(create_pipeline_pool(connection_info, Box::new(builder)))
-                        }
-                        #[cfg(not(feature = "upstream-doh3"))]
-                        {
-                            return Err(DnsError::plugin(
-                                "upstream DoH3 (HTTP/3) is not compiled into this build; \
-                                 rebuild with --features upstream-doh3",
-                            ));
-                        }
-                    } else {
-                        let builder = H2ConnectionBuilder::new(&connection_info);
-                        Box::new(create_pipeline_pool(connection_info, Box::new(builder)))
-                    }
-                }
-                #[cfg(not(feature = "upstream-doh"))]
-                ConnectionType::DoH => {
-                    return Err(DnsError::plugin(
-                        "upstream DoH is not compiled into this build; \
-                         rebuild with --features upstream-doh",
-                    ));
-                }
-            };
-            Ok(upstream)
-        } else {
-            // Domain-based upstream: use bootstrap or system DNS for resolution
-            let upstream: Box<dyn Upstream> = match &connection_info.connection_type {
-                ConnectionType::UDP => {
-                    let upstream: BootstrapUpstream<UdpConnection> =
-                        BootstrapUpstream::new(connection_info);
-                    Box::new(upstream)
-                }
-                ConnectionType::TCP => {
-                    let upstream: BootstrapUpstream<TcpConnection> =
-                        BootstrapUpstream::new(connection_info);
-                    Box::new(upstream)
-                }
-                #[cfg(feature = "upstream-dot")]
-                ConnectionType::DoT => {
-                    let upstream: BootstrapUpstream<TcpConnection> =
-                        BootstrapUpstream::new(connection_info);
-                    Box::new(upstream)
-                }
-                #[cfg(not(feature = "upstream-dot"))]
-                ConnectionType::DoT => {
-                    return Err(DnsError::plugin(
-                        "upstream DoT is not compiled into this build; \
-                         rebuild with --features upstream-dot",
-                    ));
-                }
-                #[cfg(feature = "upstream-doq")]
-                ConnectionType::DoQ => {
-                    let upstream: BootstrapUpstream<QuicConnection> =
-                        BootstrapUpstream::new(connection_info);
-                    Box::new(upstream)
-                }
-                #[cfg(not(feature = "upstream-doq"))]
-                ConnectionType::DoQ => {
-                    return Err(DnsError::plugin(
-                        "upstream DoQ is not compiled into this build; \
-                         rebuild with --features upstream-doq",
-                    ));
-                }
-                #[cfg(feature = "upstream-doh")]
-                ConnectionType::DoH => {
-                    if connection_info.enable_http3 {
-                        #[cfg(feature = "upstream-doh3")]
-                        {
-                            let upstream: BootstrapUpstream<H3Connection> =
-                                BootstrapUpstream::new(connection_info);
-                            Box::new(upstream)
-                        }
-                        #[cfg(not(feature = "upstream-doh3"))]
-                        {
-                            return Err(DnsError::plugin(
-                                "upstream DoH3 (HTTP/3) is not compiled into this build; \
-                                 rebuild with --features upstream-doh3",
-                            ));
-                        }
-                    } else {
-                        let upstream: BootstrapUpstream<H2Connection> =
-                            BootstrapUpstream::new(connection_info);
-                        Box::new(upstream)
-                    }
-                }
-                #[cfg(not(feature = "upstream-doh"))]
-                ConnectionType::DoH => {
-                    return Err(DnsError::plugin(
-                        "upstream DoH is not compiled into this build; \
-                         rebuild with --features upstream-doh",
-                    ));
-                }
-            };
-            Ok(upstream)
-        }
-    }
-
-    /// Build an upstream instance from configuration
-    pub fn with_upstream_config(upstream_config: UpstreamConfig) -> Result<Box<dyn Upstream>> {
-        let connection_info = ConnectionInfo::try_from(upstream_config)?;
-        debug!("create upstream, connection info: {:?}", connection_info);
-        Self::with_connection_info(connection_info)
-    }
-}
-
-#[inline]
-const fn pipeline_request_map_capacity() -> u16 {
-    ConnectionInfo::DEFAULT_MAX_CONNS_LOAD
-}
-
-#[inline]
-const fn reuse_request_map_capacity() -> u16 {
-    1
-}
-
-#[inline]
-fn main_pool_min_conns(connection_info: &ConnectionInfo) -> usize {
-    connection_info.min_conns_or_default()
-}
-
-#[inline]
-const fn udp_truncated_fallback_min_conns() -> usize {
-    0
-}
-
-fn create_pipeline_pool<C: Connection>(
-    connection_info: ConnectionInfo,
-    builder: Box<dyn ConnectionBuilder<C>>,
-) -> PooledUpstream<C> {
-    let timeout = connection_info.timeout;
-    let min_size = main_pool_min_conns(&connection_info);
-    PooledUpstream::<C> {
-        pool: PipelinePool::new(
-            min_size,
-            connection_info.max_conns_or_default(),
-            ConnectionInfo::DEFAULT_MAX_CONNS_LOAD,
-            connection_info.idle_timeout,
-            builder,
-            QueryTimeoutPolicy::Retire,
-            timeout,
-        ),
-        connection_info,
-    }
-}
-
-fn create_reuse_pool<C: Connection>(
-    connection_info: ConnectionInfo,
-    builder: Box<dyn ConnectionBuilder<C>>,
-) -> PooledUpstream<C> {
-    let timeout = connection_info.timeout;
-    let min_size = main_pool_min_conns(&connection_info);
-    PooledUpstream::<C> {
-        pool: ReusePool::new(
-            min_size,
-            connection_info.max_conns_or_default(),
-            connection_info.idle_timeout,
-            builder,
-            QueryTimeoutPolicy::Close,
-            timeout,
-        ),
-        connection_info,
-    }
-}
-
-/// Pooled upstream resolver implementation
-///
-/// Uses connection pooling to efficiently reuse connections for multiple
-/// queries. The pool type (pipeline or reuse) is determined during creation
-/// based on protocol capabilities and configuration.
-#[allow(unused)]
-#[derive(Debug)]
-struct PooledUpstream<C: Connection> {
-    /// Connection metadata (remote address, port, etc.)
-    connection_info: ConnectionInfo,
-    /// Connection pool for load balancing and connection reuse
-    pool: Arc<dyn ConnectionPool<C>>,
-}
-
-#[async_trait]
-impl<C: Connection> Upstream for PooledUpstream<C> {
-    /// Execute DNS query through the connection pool
-    ///
-    /// The pool handles connection selection, creation, and lifecycle
-    /// management. No additional logging here as the pool layer already
-    /// logs connection events.
-    async fn inner_query(&self, request: Message, deadline: QueryDeadline) -> Result<Message> {
-        self.pool.query(request, deadline).await
-    }
-
-    fn connection_info(&self) -> &ConnectionInfo {
-        &self.connection_info
-    }
-
-    fn handles_query_deadline(&self) -> bool {
-        true
-    }
-}
-
-/// UDP upstream with automatic TCP fallback on truncation
-///
-/// DNS over UDP has a 512-byte size limit (or EDNS extended size).
-/// When responses exceed this limit, the TC (truncated) bit is set,
-/// indicating the client should retry over TCP to get the full response.
-///
-/// This upstream automatically handles this fallback:
-/// 1. Try UDP first (fast, low overhead)
-/// 2. If truncated, automatically retry over TCP
-#[derive(Debug)]
-struct UdpTruncatedUpstream {
-    /// Connection configuration (includes timeout)
-    connection_info: ConnectionInfo,
-    /// Primary UDP connection pool (fast path)
-    main_pool: Arc<dyn ConnectionPool<UdpConnection>>,
-    /// Fallback TCP connection pool (used when UDP response is truncated)
-    fallback_pool: Arc<dyn ConnectionPool<TcpConnection>>,
-}
-
-#[async_trait]
-impl Upstream for UdpTruncatedUpstream {
-    async fn inner_query(&self, request: Message, deadline: QueryDeadline) -> Result<Message> {
-        // Try UDP first (most DNS queries fit in UDP packets)
-        let response = self.main_pool.query(request.clone(), deadline).await?;
-
-        // Check if response was truncated (TC bit set)
-        if response.truncated() {
-            // Log fallback event (only happens occasionally, minimal performance impact)
-            debug!("UDP response truncated, falling back to TCP");
-
-            // Retry over TCP to get the full response
-            self.fallback_pool.query(request, deadline).await
-        } else {
-            // UDP response was complete, return it
-            Ok(response)
-        }
-    }
-
-    fn connection_info(&self) -> &ConnectionInfo {
-        &self.connection_info
-    }
-
-    fn handles_query_deadline(&self) -> bool {
-        true
-    }
-}
-
-#[derive(Debug)]
-pub struct ConnectionBuilderFactory {
-    connection_info: ConnectionInfo,
-}
-
-impl ConnectionBuilderFactory {
-    pub fn new(connection_info: ConnectionInfo) -> Self {
-        ConnectionBuilderFactory { connection_info }
-    }
-
-    /// Build a ConnectionBuilder with the resolved IP address.
-    ///
-    /// # Safety
-    ///
-    /// This method uses `unsafe transmute` to convert concrete
-    /// ConnectionBuilder types to the generic type `C`. This is SAFE
-    /// because:
-    ///
-    /// 1. The generic parameter `C` in `DomainUpstream<C>` is determined at
-    ///    creation time based on `connection_info.connection_type`
-    /// 2. `connection_info.connection_type` is immutable and never changes at
-    ///    runtime
-    /// 3. The match ensures we always transmute the correct concrete type to
-    ///    `C`:
-    ///    - `ConnectionType::UDP` is always used with
-    ///      `DomainUpstream<UdpConnection>`
-    ///    - `ConnectionType::TCP` is always used with
-    ///      `DomainUpstream<TcpConnection>`
-    ///    - etc.
-    ///
-    /// The type invariant is established in
-    /// `UpstreamBuilder::with_upstream_config()` where `DomainUpstream<C>`
-    /// is created with the matching `C` for each ConnectionType.
-    pub fn build<C: Connection>(
-        &self,
-        ip: IpAddr,
-        request_map_capacity: u16,
-    ) -> Box<dyn ConnectionBuilder<C>> {
-        let mut info = self.connection_info.clone();
-        info.remote_ip = Some(ip);
-        match info.connection_type {
-            ConnectionType::UDP => {
-                let src: Box<dyn ConnectionBuilder<UdpConnection>> =
-                    Box::new(UdpConnectionBuilder::new(&info, request_map_capacity));
-                unsafe {
-                    std::mem::transmute::<
-                        Box<dyn ConnectionBuilder<UdpConnection>>,
-                        Box<dyn ConnectionBuilder<C>>,
-                    >(src)
-                }
-            }
-            ConnectionType::TCP => {
-                let src: Box<dyn ConnectionBuilder<TcpConnection>> =
-                    Box::new(TcpConnectionBuilder::new(&info, request_map_capacity));
-                unsafe {
-                    std::mem::transmute::<
-                        Box<dyn ConnectionBuilder<TcpConnection>>,
-                        Box<dyn ConnectionBuilder<C>>,
-                    >(src)
-                }
-            }
-            #[cfg(feature = "upstream-dot")]
-            ConnectionType::DoT => {
-                let src: Box<dyn ConnectionBuilder<TcpConnection>> =
-                    Box::new(TcpConnectionBuilder::new(&info, request_map_capacity));
-                unsafe {
-                    std::mem::transmute::<
-                        Box<dyn ConnectionBuilder<TcpConnection>>,
-                        Box<dyn ConnectionBuilder<C>>,
-                    >(src)
-                }
-            }
-            #[cfg(not(feature = "upstream-dot"))]
-            ConnectionType::DoT => {
-                unreachable!("upstream DoT branch reached but feature `upstream-dot` is disabled")
-            }
-            #[cfg(feature = "upstream-doq")]
-            ConnectionType::DoQ => {
-                let src: Box<dyn ConnectionBuilder<QuicConnection>> =
-                    Box::new(QuicConnectionBuilder::new(&info));
-                unsafe {
-                    std::mem::transmute::<
-                        Box<dyn ConnectionBuilder<QuicConnection>>,
-                        Box<dyn ConnectionBuilder<C>>,
-                    >(src)
-                }
-            }
-            #[cfg(not(feature = "upstream-doq"))]
-            ConnectionType::DoQ => {
-                // Unreachable: with_connection_info refuses DoQ when the feature
-                // is off, so a BootstrapUpstream that would call back into this
-                // builder is never constructed.
-                unreachable!("upstream DoQ branch reached but feature `upstream-doq` is disabled")
-            }
-            #[cfg(feature = "upstream-doh")]
-            ConnectionType::DoH => {
-                if info.enable_http3 {
-                    #[cfg(feature = "upstream-doh3")]
-                    {
-                        let src: Box<dyn ConnectionBuilder<H3Connection>> =
-                            Box::new(H3ConnectionBuilder::new(&info));
-                        unsafe {
-                            std::mem::transmute::<
-                                Box<dyn ConnectionBuilder<H3Connection>>,
-                                Box<dyn ConnectionBuilder<C>>,
-                            >(src)
-                        }
-                    }
-                    #[cfg(not(feature = "upstream-doh3"))]
-                    {
-                        unreachable!(
-                            "upstream DoH3 branch reached but feature `upstream-doh3` is disabled"
-                        )
-                    }
-                } else {
-                    let src: Box<dyn ConnectionBuilder<H2Connection>> =
-                        Box::new(H2ConnectionBuilder::new(&info));
-                    unsafe {
-                        std::mem::transmute::<
-                            Box<dyn ConnectionBuilder<H2Connection>>,
-                            Box<dyn ConnectionBuilder<C>>,
-                        >(src)
-                    }
-                }
-            }
-            #[cfg(not(feature = "upstream-doh"))]
-            ConnectionType::DoH => {
-                unreachable!("upstream DoH branch reached but feature `upstream-doh` is disabled")
-            }
-        }
-    }
-}
-
-/// Domain-based upstream resolver that uses bootstrap to resolve domain names
-///
-/// When the upstream server is specified as a domain name (e.g.,
-/// dns.google.com) instead of an IP address, we need to resolve it first. This
-/// creates a chicken-and-egg problem: we need DNS to resolve the DNS server's
-/// address!
-///
-/// This upstream solves it by using a bootstrap resolver:
-/// 1. Bootstrap resolver (configured with IP) resolves the domain name
-/// 2. Resolved IP is cached with TTL
-/// 3. Connection pool is created/updated when IP changes
-/// 4. DNS queries are forwarded through the pool
-///
-/// # Performance
-/// - Lock-free pool swapping using ArcSwap (no blocking on IP changes)
-/// - IP resolution is cached, not done on every query
-/// - Automatic pool refresh when TTL expires
-#[derive(Debug)]
-struct BootstrapUpstream<C: Connection> {
-    /// Upstream server domain name (for logging)
-    server_name: String,
-    /// Connection metadata (includes bootstrap config)
-    connection_info: ConnectionInfo,
-    /// Bootstrap resolver for domain name resolution
-    bootstrap: Arc<Bootstrap>,
-    /// Lock-free connection pool with current resolved IP
-    /// Tuple: (current_ip, connection_pool)
-    pool: ArcSwap<(Option<IpAddr>, Arc<dyn ConnectionPool<C>>)>,
-    /// Factory for creating connection builders when IP changes
-    builder_factory: ConnectionBuilderFactory,
-    /// Serializes cold-path pool creation after bootstrap refreshes.
-    pool_update_lock: Mutex<()>,
-}
-
-impl<C: Connection> BootstrapUpstream<C> {
-    /// Create a new domain upstream with the given connection info and optional
-    /// bootstrap server
-    fn new(connection_info: ConnectionInfo) -> Self {
-        let pool: Arc<dyn ConnectionPool<C>> = ReusePool::<C>::new(
-            0,
-            1,
-            ConnectionInfo::DEFAULT_CONN_IDLE_TIME,
-            Box::new(DummyConnectionBuilder {}),
-            QueryTimeoutPolicy::Close,
-            connection_info.timeout,
-        );
-
-        let conn_info = connection_info.clone();
-        let builder_factory = ConnectionBuilderFactory::new(conn_info.clone());
-        BootstrapUpstream {
-            server_name: connection_info.server_name.clone(),
-            bootstrap: connection_info.bootstrap.clone().unwrap(),
-            connection_info,
-            pool: ArcSwap::from_pointee((None, pool)),
-            builder_factory,
-            pool_update_lock: Mutex::new(()),
-        }
-    }
-
-    /// Initialize or refresh the connection pool with the resolved IP
-    ///
-    /// This method handles:
-    /// - Initial pool creation on first query
-    /// - IP change detection and pool refresh
-    /// - Lock-free pool updates using ArcSwap
-    ///
-    /// # Performance
-    /// - Fast path: cached bootstrap IP + single atomic pool load when IP
-    ///   hasn't changed
-    /// - Pool recreation only happens on IP change (rare)
-    /// - Cold-path pool recreation is serialized to avoid duplicate pool builds
-    async fn init_pool_if_needed(&self, deadline: QueryDeadline) -> Result<()> {
-        // Fast path: atomically load current pool state (lock-free)
-        let guard = &(*self.pool.load());
-        let pool_ip = guard.0;
-
-        // Resolve domain name via bootstrap (cached in Bootstrap with TTL)
-        let ip = match self.bootstrap.get_with_deadline(deadline).await {
-            Ok(value) => value,
-            Err(value) => return Err(value),
-        };
-
-        // Check if IP has changed since last resolution
-        if let Some(current_ip) = pool_ip
-            && current_ip == ip
-        {
-            // IP unchanged, continue using current pool (hot path)
-            return Ok(());
-        }
-
-        let _update_guard = match deadline.run(self.pool_update_lock.lock()).await {
-            DeadlineOutcome::Completed(guard) => guard,
-            DeadlineOutcome::Expired => return Err(deadline.timeout_error()),
-        };
-        let guard = &(*self.pool.load());
-        let pool_ip = guard.0;
-
-        if let Some(current_ip) = pool_ip {
-            if current_ip == ip {
-                // Another waiter already refreshed the pool while we were
-                // waiting for the cold-path update lock.
-                return Ok(());
-            }
-
-            // IP changed - log the change (rare event, acceptable overhead)
-            info!(
-                server = %self.server_name,
-                old_ip = %current_ip,
-                new_ip = %ip,
-                "Upstream IP address changed, refreshing connection pool"
-            );
-        } else {
-            // First initialization
-            info!(
-                server = %self.server_name,
-                ip = %ip,
-                "Initializing connection pool for domain-based upstream"
-            );
-        }
-
-        // Create new connection builder with the resolved IP
-        let request_map_capacity = match self.connection_info.connection_type {
-            ConnectionType::UDP => pipeline_request_map_capacity(),
-            ConnectionType::TCP | ConnectionType::DoT => {
-                if self.connection_info.enable_pipeline.unwrap_or(false) {
-                    pipeline_request_map_capacity()
-                } else {
-                    reuse_request_map_capacity()
-                }
-            }
-            ConnectionType::DoQ | ConnectionType::DoH => reuse_request_map_capacity(),
-        };
-
-        let builder: Box<dyn ConnectionBuilder<C>> =
-            self.builder_factory.build(ip, request_map_capacity);
-
-        // Create appropriate pool type based on protocol
-        let new_pool: Arc<dyn ConnectionPool<C>> = match self.connection_info.connection_type {
-            ConnectionType::UDP => PipelinePool::new(
-                main_pool_min_conns(&self.connection_info),
-                self.connection_info.max_conns_or_default(),
-                ConnectionInfo::DEFAULT_MAX_CONNS_LOAD,
-                self.connection_info.idle_timeout,
-                builder,
-                QueryTimeoutPolicy::Reuse,
-                self.connection_info.timeout,
-            ),
-            ConnectionType::TCP | ConnectionType::DoT => {
-                if self.connection_info.enable_pipeline.unwrap_or(false) {
-                    PipelinePool::new(
-                        main_pool_min_conns(&self.connection_info),
-                        self.connection_info.max_conns_or_default(),
-                        ConnectionInfo::DEFAULT_MAX_CONNS_LOAD,
-                        self.connection_info.idle_timeout,
-                        builder,
-                        QueryTimeoutPolicy::Retire,
-                        self.connection_info.timeout,
-                    )
-                } else {
-                    ReusePool::new(
-                        main_pool_min_conns(&self.connection_info),
-                        self.connection_info.max_conns_or_default(),
-                        self.connection_info.idle_timeout,
-                        builder,
-                        QueryTimeoutPolicy::Close,
-                        self.connection_info.timeout,
-                    )
-                }
-            }
-            ConnectionType::DoQ | ConnectionType::DoH => PipelinePool::new(
-                main_pool_min_conns(&self.connection_info),
-                self.connection_info.max_conns_or_default(),
-                ConnectionInfo::DEFAULT_MAX_CONNS_LOAD,
-                self.connection_info.idle_timeout,
-                builder,
-                QueryTimeoutPolicy::Retire,
-                self.connection_info.timeout,
-            ),
-        };
-
-        // Atomically swap to new pool (lock-free, readers see old or new pool
-        // consistently)
-        self.pool.swap(Arc::from((Some(ip), new_pool)));
-
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl<C: Connection> Upstream for BootstrapUpstream<C> {
-    /// Execute DNS query through bootstrap-resolved upstream
-    ///
-    /// # Process
-    /// 1. Resolve domain name to IP (cached with TTL in bootstrap)
-    /// 2. Initialize/refresh pool if IP changed
-    /// 3. Forward query through the pool
-    ///
-    /// # Performance
-    /// - Hot path: pool already initialized, just forward query
-    /// - Cold path: bootstrap resolution + pool creation (first query only)
-    /// - IP change: new pool creation (rare, based on DNS TTL)
-    async fn inner_query(&self, request: Message, deadline: QueryDeadline) -> Result<Message> {
-        // Ensure connection pool is initialized with current IP
-        // Fast path: just checks atomic, no allocation
-        // Slow path: resolves DNS + creates pool (only on first query or IP change)
-        self.init_pool_if_needed(deadline).await?;
-
-        // Get current connection pool (lock-free atomic load)
-        let pool = &*self.pool.load();
-
-        // Forward query through the pool
-        pool.1.query(request, deadline).await
-    }
-
-    fn connection_info(&self) -> &ConnectionInfo {
-        &self.connection_info
-    }
-
-    fn handles_query_deadline(&self) -> bool {
-        true
-    }
-}
-
-/// Dummy connection builder for initial empty pool
-///
-/// This is used as a placeholder before the first DNS resolution completes.
-/// Any attempt to create a connection will fail with an error.
-#[derive(Debug)]
-struct DummyConnectionBuilder {}
-
-#[async_trait]
-impl<C: Connection> ConnectionBuilder<C> for DummyConnectionBuilder {
-    async fn create_connection(&self, _conn_id: u16, _deadline: QueryDeadline) -> Result<Arc<C>> {
-        Err(DnsError::protocol(
-            "DummyConnectionBuilder cannot create connections (pool not yet initialized)",
-        ))
-    }
-}
-
-/// Parse SOCKS5 proxy configuration from string
-///
-/// Supports two formats:
-/// - "host:port" - SOCKS5 without authentication
-/// - "username:password@host:port" - SOCKS5 with authentication
-///
-/// If host is a domain name, it will be resolved using system DNS.
-///
-/// # Arguments
-/// * `socks5_str` - SOCKS5 proxy string in one of the supported formats
-///
-/// # Returns
-/// - `Some(Socks5Opt)` if parsing and resolution succeed
-/// - `None` if parsing fails or hostname resolution fails
-///
-/// # Examples
-/// ```text
-/// // Without auth
-/// parse_socks5_opt("127.0.0.1:1080")
-/// parse_socks5_opt("proxy.example.com:1080")
-///
-/// // With auth
-/// parse_socks5_opt("user:pass@127.0.0.1:1080")
-/// parse_socks5_opt("user:pass@proxy.example.com:1080")
-/// ```
-fn parse_socks5_opt_with_resolver<F>(socks5_str: &str, mut resolve_host: F) -> Option<Socks5Opt>
-where
-    F: FnMut(&str) -> Result<IpAddr>,
-{
-    // Split by '@' to separate auth from host:port
-    let (username, password, host_port) = if let Some(at_pos) = socks5_str.rfind('@') {
-        // Format: username:password@host:port
-        let auth_part = &socks5_str[..at_pos];
-        let host_part = &socks5_str[at_pos + 1..];
-
-        // Split auth by ':'
-        if let Some(colon_pos) = auth_part.find(':') {
-            let username = auth_part[..colon_pos].to_string();
-            let password = auth_part[colon_pos + 1..].to_string();
-            (Some(username), Some(password), host_part)
-        } else {
-            warn!(
-                "Invalid SOCKS5 auth format (expected username:password): {}",
-                socks5_str
-            );
-            return None;
-        }
-    } else {
-        // Format: host:port (no auth)
-        (None, None, socks5_str)
-    };
-
-    // Parse host:port - use last colon to split
-    let (mut host, port) = match host_port.rfind(':') {
-        Some(colon_pos) => {
-            let host = &host_port[..colon_pos];
-            let port_str = &host_port[colon_pos + 1..];
-
-            match port_str.parse::<u16>() {
-                Ok(port) => (host, port),
-                Err(_) => {
-                    warn!("Invalid SOCKS5 port: {}", port_str);
-                    return None;
-                }
-            }
-        }
-        None => {
-            warn!("Invalid SOCKS5 format (expected host:port): {}", host_port);
-            return None;
-        }
-    };
-
-    // Remove IPv6 brackets if present: [::1] -> ::1
-    if host.starts_with('[') && host.ends_with(']') {
-        host = &host[1..host.len() - 1];
-    }
-
-    // Resolve host to IP address
-    let ip_addr = if let Ok(ip) = IpAddr::from_str(host) {
-        // Already an IP address
-        ip
-    } else {
-        // It's a hostname, resolve it
-        match resolve_host(host) {
-            Ok(ip) => ip,
-            Err(e) => {
-                warn!("Failed to resolve SOCKS5 hostname '{}': {}", host, e);
-                return None;
-            }
-        }
-    };
-
-    Some(Socks5Opt {
-        username,
-        password,
-        socket_addr: SocketAddr::new(ip_addr, port),
-    })
-}
-
-pub(crate) fn parse_socks5_opt(socks5_str: &str) -> Option<Socks5Opt> {
-    parse_socks5_opt_with_resolver(socks5_str, try_lookup_server_name)
-}
-
+pub use builder::UpstreamBuilder;
+pub use config::{ConnectionInfo, ConnectionType, UpstreamConfig};
 #[cfg(feature = "_http-client")]
-pub(crate) async fn connect_tcp_stream(
-    remote_ip: Option<IpAddr>,
-    server_name: String,
-    port: u16,
-    socks5_opt: Option<Socks5Opt>,
-) -> Result<TcpStream> {
-    utils::connect_stream(remote_ip, server_name, port, None, None, socks5_opt).await
-}
+pub(crate) use dial::connect_tcp_stream;
+pub(crate) use pool::Connection;
+pub use pool::QueryDeadline;
+pub use resolver::Upstream;
 
 #[cfg(test)]
-mod tests {
-    use std::sync::atomic::{AtomicBool, AtomicU16, AtomicU64, Ordering};
-
-    use super::*;
-
-    #[derive(Debug)]
-    struct SlowUpstream {
-        connection_info: ConnectionInfo,
-    }
-
-    #[async_trait]
-    impl Upstream for SlowUpstream {
-        async fn inner_query(&self, request: Message, _deadline: QueryDeadline) -> Result<Message> {
-            tokio::time::sleep(Duration::from_secs(60)).await;
-            Ok(request)
-        }
-
-        fn connection_info(&self) -> &ConnectionInfo {
-            &self.connection_info
-        }
-    }
-
-    #[derive(Debug)]
-    struct NoopConnection {
-        available: AtomicBool,
-        using_count: AtomicU16,
-        last_used: AtomicU64,
-    }
-
-    #[async_trait]
-    impl Connection for NoopConnection {
-        fn close(&self) {
-            self.available.store(false, Ordering::Relaxed);
-        }
-
-        async fn query(&self, request: Message, _deadline: QueryDeadline) -> Result<Message> {
-            Ok(request)
-        }
-
-        fn using_count(&self) -> u16 {
-            self.using_count.load(Ordering::Relaxed)
-        }
-
-        fn available(&self) -> bool {
-            self.available.load(Ordering::Relaxed)
-        }
-
-        fn last_used(&self) -> u64 {
-            self.last_used.load(Ordering::Relaxed)
-        }
-    }
-
-    #[derive(Debug)]
-    struct NoopConnectionBuilder;
-
-    #[async_trait]
-    impl ConnectionBuilder<NoopConnection> for NoopConnectionBuilder {
-        async fn create_connection(
-            &self,
-            _conn_id: u16,
-            _deadline: QueryDeadline,
-        ) -> Result<Arc<NoopConnection>> {
-            Ok(Arc::new(NoopConnection {
-                available: AtomicBool::new(true),
-                using_count: AtomicU16::new(0),
-                last_used: AtomicU64::new(crate::core::app_clock::AppClock::elapsed_millis()),
-            }))
-        }
-    }
-
-    #[derive(Debug)]
-    struct DeadlineHandlingPool {
-        handled_timeout: Arc<AtomicBool>,
-    }
-
-    #[async_trait]
-    impl ConnectionPool<NoopConnection> for DeadlineHandlingPool {
-        async fn query(&self, _request: Message, deadline: QueryDeadline) -> Result<Message> {
-            let Some(remaining) = deadline.remaining() else {
-                self.handled_timeout.store(true, Ordering::Relaxed);
-                return Err(deadline.timeout_error());
-            };
-            tokio::time::sleep(remaining + Duration::from_millis(20)).await;
-            self.handled_timeout.store(true, Ordering::Relaxed);
-            Err(deadline.timeout_error())
-        }
-
-        async fn maintain(&self) {}
-
-        fn configured_min_size(&self) -> usize {
-            0
-        }
-    }
-
-    fn make_upstream_config(addr: &str) -> UpstreamConfig {
-        UpstreamConfig {
-            tag: None,
-            addr: addr.to_string(),
-            dial_addr: None,
-            port: None,
-            bootstrap: None,
-            bootstrap_version: None,
-            socks5: None,
-            idle_timeout: None,
-            max_conns: None,
-            min_conns: None,
-            insecure_skip_verify: None,
-            timeout: None,
-            enable_pipeline: None,
-            enable_http3: None,
-            so_mark: None,
-            bind_to_device: None,
-        }
-    }
-
-    #[test]
-    fn test_helper_scheme_tcp_pipeline_forces_pipeline() {
-        let mut cfg = make_upstream_config("tcp+pipeline://1.1.1.1");
-        cfg.enable_pipeline = Some(false);
-        let info = ConnectionInfo::try_from(cfg).expect("helper scheme should be accepted");
-        assert_eq!(info.connection_type, ConnectionType::TCP);
-        assert_eq!(info.enable_pipeline, Some(true));
-    }
-
-    #[test]
-    fn test_helper_scheme_h3_forces_http3() {
-        let mut cfg = make_upstream_config("h3://dns.google/dns-query");
-        cfg.enable_http3 = Some(false);
-        let info = ConnectionInfo::try_from(cfg).expect("helper scheme should be accepted");
-        assert_eq!(info.connection_type, ConnectionType::DoH);
-        assert!(info.enable_http3);
-    }
-
-    #[test]
-    fn test_connection_info_defers_domain_resolution() {
-        let info = ConnectionInfo::with_addr("tls://dns.example.invalid:853")
-            .expect("domain upstream should parse without DNS resolution");
-        assert_eq!(info.server_name, "dns.example.invalid");
-        assert!(info.remote_ip.is_none());
-
-        let info = ConnectionInfo::try_from(make_upstream_config(
-            "https://resolver.example.invalid/dns-query",
-        ))
-        .expect("domain upstream config should parse without DNS resolution");
-        assert_eq!(info.server_name, "resolver.example.invalid");
-        assert!(info.remote_ip.is_none());
-    }
-
-    #[test]
-    fn test_connection_info_uses_dial_addr_for_domain() {
-        let mut cfg = make_upstream_config("tls://dns.example.invalid:853");
-        cfg.dial_addr = Some(IpAddr::from_str("203.0.113.53").unwrap());
-
-        let info = ConnectionInfo::try_from(cfg).expect("upstream config should parse");
-        assert_eq!(info.server_name, "dns.example.invalid");
-        assert_eq!(
-            info.remote_ip,
-            Some(IpAddr::from_str("203.0.113.53").unwrap())
-        );
-    }
-
-    #[test]
-    fn test_connection_info_dial_addr_takes_precedence_over_bootstrap() {
-        let mut cfg = make_upstream_config("tls://dns.example.invalid:853");
-        cfg.dial_addr = Some(IpAddr::from_str("203.0.113.53").unwrap());
-        cfg.bootstrap = Some("8.8.8.8:53".to_string());
-
-        let info = ConnectionInfo::try_from(cfg).expect("upstream config should parse");
-        assert_eq!(
-            info.remote_ip,
-            Some(IpAddr::from_str("203.0.113.53").unwrap())
-        );
-        assert!(info.bootstrap.is_none());
-    }
-
-    #[test]
-    fn test_connection_info_rejects_invalid_bootstrap_version() {
-        let mut cfg = make_upstream_config("tls://dns.example.invalid:853");
-        cfg.bootstrap = Some("8.8.8.8:53".to_string());
-        cfg.bootstrap_version = Some(5);
-
-        let err = ConnectionInfo::try_from(cfg).expect_err("invalid bootstrap_version should fail");
-
-        assert!(
-            err.to_string().contains("bootstrap_version must be 4 or 6"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn test_max_conns_is_preserved() {
-        let mut cfg = make_upstream_config("8.8.8.8");
-        cfg.max_conns = Some(999);
-        let info = ConnectionInfo::try_from(cfg).expect("upstream config should parse");
-        assert_eq!(info.max_conns, Some(999));
-    }
-
-    #[test]
-    fn test_max_conns_rejects_zero() {
-        let mut cfg = make_upstream_config("8.8.8.8");
-        cfg.max_conns = Some(0);
-
-        let err = ConnectionInfo::try_from(cfg).expect_err("zero max_conns should be rejected");
-
-        assert!(
-            err.to_string().contains("max_conns must be greater than 0"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn test_max_conns_rejects_excessive_value() {
-        let mut cfg = make_upstream_config("8.8.8.8");
-        cfg.max_conns = Some(ConnectionInfo::MAX_CONFIGURED_CONNS_SIZE + 1);
-
-        let err =
-            ConnectionInfo::try_from(cfg).expect_err("excessive max_conns should be rejected");
-
-        assert!(
-            err.to_string().contains("max_conns must be <= 4096"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn test_min_conns_is_preserved() {
-        let mut cfg = make_upstream_config("8.8.8.8");
-        cfg.min_conns = Some(3);
-
-        let info = ConnectionInfo::try_from(cfg).expect("upstream config should parse");
-
-        assert_eq!(info.min_conns, Some(3));
-    }
-
-    #[test]
-    fn test_min_conns_allows_zero() {
-        let mut cfg = make_upstream_config("8.8.8.8");
-        cfg.min_conns = Some(0);
-
-        let info = ConnectionInfo::try_from(cfg).expect("zero min_conns should be accepted");
-
-        assert_eq!(info.min_conns, Some(0));
-        assert_eq!(info.min_conns_or_default(), 0);
-    }
-
-    #[test]
-    fn test_min_conns_rejects_excessive_value() {
-        let mut cfg = make_upstream_config("8.8.8.8");
-        cfg.max_conns = Some(ConnectionInfo::MAX_CONFIGURED_CONNS_SIZE);
-        cfg.min_conns = Some(ConnectionInfo::MAX_CONFIGURED_CONNS_SIZE + 1);
-
-        let err =
-            ConnectionInfo::try_from(cfg).expect_err("excessive min_conns should be rejected");
-
-        assert!(
-            err.to_string().contains("min_conns must be <= 4096"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn test_min_conns_rejects_value_above_configured_max_conns() {
-        let mut cfg = make_upstream_config("8.8.8.8");
-        cfg.max_conns = Some(2);
-        cfg.min_conns = Some(3);
-
-        let err = ConnectionInfo::try_from(cfg)
-            .expect_err("min_conns above configured max_conns should be rejected");
-
-        assert!(
-            err.to_string().contains("min_conns must be <= max_conns"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn test_min_conns_rejects_value_above_default_max_conns() {
-        let mut cfg = make_upstream_config("8.8.8.8");
-        cfg.min_conns = Some(ConnectionInfo::DEFAULT_MAX_CONNS_SIZE + 1);
-
-        let err = ConnectionInfo::try_from(cfg)
-            .expect_err("min_conns above default max_conns should be rejected");
-
-        assert!(err.to_string().contains("effective max_conns: 64"), "{err}");
-    }
-
-    #[tokio::test]
-    async fn test_pipeline_pool_uses_configured_min_conns() {
-        let mut info = ConnectionInfo::with_addr("tcp://127.0.0.1").expect("upstream should parse");
-        info.max_conns = Some(4);
-        info.min_conns = Some(2);
-
-        let upstream =
-            create_pipeline_pool::<NoopConnection>(info, Box::new(NoopConnectionBuilder));
-
-        assert_eq!(upstream.pool.configured_min_size(), 2);
-    }
-
-    #[tokio::test]
-    async fn test_reuse_pool_uses_configured_min_conns() {
-        let mut info = ConnectionInfo::with_addr("tcp://127.0.0.1").expect("upstream should parse");
-        info.max_conns = Some(4);
-        info.min_conns = Some(2);
-
-        let upstream = create_reuse_pool::<NoopConnection>(info, Box::new(NoopConnectionBuilder));
-
-        assert_eq!(upstream.pool.configured_min_size(), 2);
-    }
-
-    #[test]
-    fn test_udp_truncated_fallback_keeps_zero_min_conns() {
-        let mut info = ConnectionInfo::with_addr("udp://127.0.0.1").expect("upstream should parse");
-        info.min_conns = Some(2);
-
-        assert_eq!(main_pool_min_conns(&info), 2);
-        assert_eq!(udp_truncated_fallback_min_conns(), 0);
-    }
-
-    #[tokio::test]
-    async fn test_query_wraps_custom_upstream_in_deadline() {
-        crate::core::app_clock::AppClock::start();
-        let mut connection_info =
-            ConnectionInfo::with_addr("udp://127.0.0.1").expect("upstream should parse");
-        connection_info.timeout = Duration::from_millis(10);
-        let upstream = SlowUpstream { connection_info };
-
-        let result = upstream.query(Message::new()).await;
-
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn test_pooled_upstream_lets_pool_handle_deadline_expiry() {
-        crate::core::app_clock::AppClock::start();
-        let handled_timeout = Arc::new(AtomicBool::new(false));
-        let mut connection_info =
-            ConnectionInfo::with_addr("tcp://127.0.0.1").expect("upstream should parse");
-        connection_info.timeout = Duration::from_millis(10);
-        let upstream = PooledUpstream::<NoopConnection> {
-            connection_info,
-            pool: Arc::new(DeadlineHandlingPool {
-                handled_timeout: handled_timeout.clone(),
-            }),
-        };
-
-        let result = upstream.query(Message::new()).await;
-
-        assert!(result.is_err());
-        assert!(handled_timeout.load(Ordering::Relaxed));
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_ip_without_auth() {
-        // Test parsing IP address without authentication
-        let result = parse_socks5_opt("127.0.0.1:1080");
-        assert!(result.is_some());
-
-        let opt = result.unwrap();
-        assert!(opt.username.is_none());
-        assert!(opt.password.is_none());
-        assert_eq!(opt.socket_addr.ip(), IpAddr::from_str("127.0.0.1").unwrap());
-        assert_eq!(opt.socket_addr.port(), 1080);
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_ip_with_auth() {
-        // Test parsing IP address with authentication
-        let result = parse_socks5_opt("myuser:mypass@192.168.1.100:8080");
-        assert!(result.is_some());
-
-        let opt = result.unwrap();
-        assert_eq!(opt.username, Some("myuser".to_string()));
-        assert_eq!(opt.password, Some("mypass".to_string()));
-        assert_eq!(
-            opt.socket_addr.ip(),
-            IpAddr::from_str("192.168.1.100").unwrap()
-        );
-        assert_eq!(opt.socket_addr.port(), 8080);
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_ipv6_without_auth() {
-        // Test parsing IPv6 address without authentication
-        let result = parse_socks5_opt("[::1]:1080");
-        assert!(result.is_some());
-
-        let opt = result.unwrap();
-        assert!(opt.username.is_none());
-        assert!(opt.password.is_none());
-        assert_eq!(opt.socket_addr.ip(), IpAddr::from_str("::1").unwrap());
-        assert_eq!(opt.socket_addr.port(), 1080);
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_ipv6_with_auth() {
-        // Test parsing IPv6 address with authentication
-        let result = parse_socks5_opt("user:pass@[2001:db8::1]:8080");
-        assert!(result.is_some());
-
-        let opt = result.unwrap();
-        assert_eq!(opt.username, Some("user".to_string()));
-        assert_eq!(opt.password, Some("pass".to_string()));
-        assert_eq!(
-            opt.socket_addr.ip(),
-            IpAddr::from_str("2001:db8::1").unwrap()
-        );
-        assert_eq!(opt.socket_addr.port(), 8080);
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_ipv6_full_address() {
-        // Test parsing full IPv6 address
-        let result = parse_socks5_opt("[fe80::1234:5678:90ab:cdef]:9050");
-        assert!(result.is_some());
-
-        let opt = result.unwrap();
-        assert_eq!(
-            opt.socket_addr.ip(),
-            IpAddr::from_str("fe80::1234:5678:90ab:cdef").unwrap()
-        );
-        assert_eq!(opt.socket_addr.port(), 9050);
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_ipv6_missing_bracket() {
-        // Test IPv6 without brackets - this actually succeeds for simple cases like ::1
-        // because rfind(':') correctly splits "::1:1080" into "::1" and "1080"
-        // However, brackets are still RECOMMENDED for clarity and standards compliance
-        let result = parse_socks5_opt("::1:1080");
-        assert!(result.is_some());
-
-        let opt = result.unwrap();
-        assert_eq!(opt.socket_addr.ip(), IpAddr::from_str("::1").unwrap());
-        assert_eq!(opt.socket_addr.port(), 1080);
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_ipv6_missing_port() {
-        // Test IPv6 with brackets but no port
-        let result = parse_socks5_opt("[::1]");
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_ipv6_unclosed_bracket() {
-        // Test IPv6 with unclosed bracket
-        let result = parse_socks5_opt("[::1:1080");
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_invalid_port() {
-        // Test invalid port number
-        let result = parse_socks5_opt("127.0.0.1:invalid");
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_missing_port() {
-        // Test missing port
-        let result = parse_socks5_opt("127.0.0.1");
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_invalid_auth_format() {
-        // Test invalid auth format (missing password)
-        let result = parse_socks5_opt("myuser@127.0.0.1:1080");
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_password_with_colon() {
-        // Test password containing colon
-        let result = parse_socks5_opt("user:pass:word@127.0.0.1:1080");
-        assert!(result.is_some());
-
-        let opt = result.unwrap();
-        assert_eq!(opt.username, Some("user".to_string()));
-        assert_eq!(opt.password, Some("pass:word".to_string()));
-        assert_eq!(opt.socket_addr.port(), 1080);
-    }
-
-    #[test]
-    fn test_parse_socks5_opt_hostname_uses_resolver() {
-        let result = parse_socks5_opt_with_resolver("localhost:1080", |host| {
-            assert_eq!(host, "localhost");
-            Ok(IpAddr::from_str("127.0.0.1").unwrap())
-        });
-        assert!(result.is_some());
-
-        let opt = result.unwrap();
-        assert!(opt.username.is_none());
-        assert!(opt.password.is_none());
-        assert_eq!(opt.socket_addr.port(), 1080);
-        assert_eq!(opt.socket_addr.ip(), IpAddr::from_str("127.0.0.1").unwrap());
-    }
-}
+mod tests;
