+++ src/network/upstream/pool/mod.rs
@@ -36,7 +36,7 @@
 
 use std::fmt::Debug;
 use std::future::Future;
-use std::sync::atomic::{AtomicU16, AtomicU64, Ordering};
+use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, Mutex, Weak};
 use std::time::Duration;
 
@@ -48,16 +48,6 @@ use crate::core::error::{DnsError, Result};
 use crate::core::task_center;
 use crate::proto::Message;
 
-mod request_map;
-
-#[cfg(feature = "upstream-doh")]
-pub(crate) mod conn_h2;
-#[cfg(feature = "upstream-doh3")]
-pub(crate) mod conn_h3;
-#[cfg(feature = "upstream-doq")]
-pub(crate) mod conn_quic;
-pub(crate) mod conn_tcp;
-pub(crate) mod conn_udp;
 pub(crate) mod pool_pipeline;
 pub(crate) mod pool_reuse;
 
@@ -219,20 +209,6 @@ pub trait ConnectionPool<C: Connection>: Send + Sync + Debug + 'static {
 pub trait ManagedMaintenanceTask {
     fn maintenance_task_id(&self) -> &Mutex<Option<u64>>;
     fn maintenance_task_name(&self) -> String;
-}
-
-/// RAII guard that decrements a connection's in-flight query counter on drop.
-///
-/// Ensures `using_count` is always decremented even when the query future is
-/// cancelled by an outer timeout, preventing the pool from permanently
-/// deadlocking due to a leaked counter.
-#[allow(dead_code)]
-pub(crate) struct UsingCountGuard<'a>(pub(crate) &'a AtomicU16);
-
-impl Drop for UsingCountGuard<'_> {
-    fn drop(&mut self) {
-        self.0.fetch_sub(1, Ordering::Relaxed);
-    }
 }
 
 /// Maintenance interval for pool cleanup
